Move secondary pane surface contracts
