+++ docs/DECISIONS.md
@@ -148,6 +148,7 @@
 - Settings surface contracts live in `QuillCodeSettingsSurface.swift`. `WorkspaceSettingsSurface`, `WorkspaceSettingsUpdate`, Computer Use requirement rows, TrustedRouter sign-in copy, Computer Use permission labels, and compatibility decoding stay beside the settings contract; `WorkspaceSurface.swift` should only carry the aggregate `settings` payload and pass current runtime state into it.
 - Terminal and browser utility panes live in `QuillCodeTerminalBrowserPaneView.swift`. Terminal history rows, browser navigation, page snapshots, comments, and pane-local draft state should evolve there; `WorkspaceSwiftUIView.swift` should only decide when the panes are visible and route the resulting actions.
 - Secondary utility panes live in `QuillCodeSecondaryPanesView.swift`. Extensions, Memories, Automations, their count/status pills, empty states, and compact cards should evolve there; `WorkspaceSwiftUIView.swift` should only decide pane placement and route commands.
+- Secondary pane surface contracts live in `QuillCodeSecondaryPaneSurface.swift`. `WorkspaceExtensionsSurface`, `WorkspaceMemoriesSurface`, `WorkspaceAutomationsSurface`, extension manifest rows, memory rows, and automation workflow rows should evolve beside the secondary-pane native and HTML renderers; `WorkspaceSurface.swift` should only carry the aggregate slots and assemble them from model state.
 - Workspace modal surfaces live in `QuillCodeWorkspaceDialogs.swift`. Command palette, keyboard shortcuts, chat search, rename sheets, and worktree create/remove sheets should evolve there; `WorkspaceSwiftUIView.swift` should only decide when a sheet is presented and route completed actions.
 - Desktop app bootstrap lives in `QuillCodeDesktopApp.swift` and should remain declarative. Native commands, menu-bar layout, OAuth loopback capture, browser fetching, automation notifications, and controller orchestration live in separate desktop files. Parity gates intentionally scan the whole `Sources/quill-code-desktop` folder so extraction can continue without pushing implementation details back into scene composition.
 - Desktop controllers should route UI and apply workspace state, not own protocol flows or cancellable task bookkeeping. TrustedRouter sign-in lives in `QuillCodeDesktopSignInCoordinator`, and send/terminal/browser/automation task slots live in `QuillCodeDesktopTaskCoordinator`. Parity gates check those boundaries so OAuth exchange and raw task IDs do not drift back into `QuillCodeDesktopController.swift`.
